+++ metadata.rb
@@ -4,7 +4,7 @@ maintainer_email 'broderson@gmail.com'
 license 'MIT'
 description 'Installs/Configures d-base'
 long_description 'Installs/Configures d-base'
-version '0.6.0'
+version '0.6.2'
 
 depends 'chef-client'
 depends 'resolver'
@@ -28,10 +28,12 @@ depends 'qgis'
 
 depends 'docker'
 
-depends 'ruby_install', '~> 1.1.0'      # special branch
-depends 'chruby_install', '~> 0.1.3'    # special branch
+#depends 'ruby_install', '~> 1.1.0'      # special branch
+#depends 'chruby_install', '~> 0.1.3'    # special branch
 
 depends 'n1nj4-users'
 
 depends 'nodejs'
 depends 'steam', '~> 2.2.1'             # special branch
+
+depends 'ruby_rbenv'