+++ recipes/_ruby.rb
@@ -4,6 +4,6 @@
 #
 # Copyright:: 2017, The Authors, All Rights Reserved.
 
-
-#include_recipe 'ruby_install'
-include_recipe 'chruby_install'
+rbenv_system_install 'system' do
+  update_rbenv false
+end